Refactor SQL rewrite pipeline
