--- src/webserver/database/sql/rewrite.rs
@@ -1,1177 +1,0 @@
-//! Compiler pass that partitions parsed SQL between the database and `SQLPage`.
-//!
-//! `SQLPage` files contain expressions owned by two runtimes: the configured
-//! database handles ordinary SQL, while variables and `sqlpage.*` functions are
-//! evaluated by `SQLPage`. This module makes that boundary explicit. Values
-//! needed before a database statement become typed bindings; computed
-//! projections run once per returned row, with any database-owned dependencies
-//! appended to the SQL projection as a private trailing column suffix. Queries
-//! that require no database work become single-row `SQLPage` plans instead. The
-//! distinct standalone and row expression types prevent pre-query work from
-//! depending on a row that does not exist yet.
-//!
-//! Rewriting also renders backend-appropriate placeholders and casts, preserves
-//! selected DBMS semantics for operations `SQLPage` emulates, records which
-//! returned values need JSON decoding, and rejects grouping, ordering, or other
-//! relational clauses that would incorrectly depend on post-database computed
-//! columns. Its output is the immutable query representation consumed by
-//! `execute_queries`; no request values are resolved and no functions are
-//! executed here.
-
-use std::ops::ControlFlow;
-use std::str::FromStr as _;
-
-use anyhow::{Context as _, anyhow};
-use serde_json::Value as JsonValue;
-use sqlparser::ast::{
-    BinaryOperator, CastKind, CharacterLength, DataType, Expr as SqlExpr, Function, FunctionArg,
-    FunctionArgExpr, FunctionArgumentList, FunctionArguments, GroupByExpr, Ident, ObjectName,
-    ObjectNamePart, OrderByKind, SelectItem, SetExpr, Statement as SqlStatement, Value,
-    ValueWithSpan, VisitMut, VisitorMut,
-};
-use sqlparser::tokenizer::Span;
-
-use super::dialect::{PlaceholderStyle, placeholder_style};
-use super::statement::{
-    DatabaseQuery, OutputColumn, Query, QueryBody, SingleRowQuery, SourceLocation, SourceSpan,
-};
-use super::{extract_json_columns, is_json_expression, is_sqlpage_func};
-use crate::webserver::database::sqlpage_expr::{
-    NoRowInput, RowExpr, RowInputId, SqlPageExpr, StandaloneExpr, VariableRef, VariableSource,
-};
-use crate::webserver::database::sqlpage_functions::functions::SqlPageFunctionName;
-use crate::webserver::database::{DbInfo, SupportedDatabase};
-use sqlx::any::AnyKind;
-
-const SQLPAGE_INPUT_PREFIX: &str = "__sqlpage_input_";
-
-/// State used while partitioning source expressions between `SQLPage` and the database.
-struct QueryPartitioner<'a> {
-    database: &'a DbInfo,
-    row_input_json: Vec<bool>,
-    private_projection: Vec<SelectItem>,
-}
-
-/// State used after partitioning to lower `SQLPage` values into database bindings.
-struct DatabaseLowerer<'a> {
-    database: &'a DbInfo,
-    bindings: Vec<StandaloneExpr>,
-    error: Option<anyhow::Error>,
-}
-
-struct ComputedAliasFinder<'a> {
-    computed_columns: &'a [OutputColumn<RowExpr>],
-}
-
-impl sqlparser::ast::Visitor for ComputedAliasFinder<'_> {
-    type Break = ();
-
-    fn pre_visit_expr(&mut self, expression: &SqlExpr) -> ControlFlow<Self::Break> {
-        let SqlExpr::Identifier(identifier) = expression else {
-            return ControlFlow::Continue(());
-        };
-        if self
-            .computed_columns
-            .iter()
-            .any(|column| identifier.value.eq_ignore_ascii_case(&column.name))
-        {
-            return ControlFlow::Break(());
-        }
-        ControlFlow::Continue(())
-    }
-}
-
-/// Result of partitioning one projected expression.
-// Keeping the owned AST inline avoids one heap allocation for every ordinary
-// projected expression. The enum is short-lived inside the rewriter.
-#[allow(clippy::large_enum_variant)]
-enum PartitionedProjection {
-    Database(SqlExpr),
-    PerRow(RowExpr),
-}
-
-/// A partitioned projection after its database fragments have been lowered.
-#[allow(clippy::large_enum_variant)]
-enum LoweredProjection {
-    Database(SqlExpr),
-    PerRow(RowExpr),
-}
-
-/// Defines how a SQL expression crossing into a SQLPage-owned expression is
-/// represented at a particular evaluation site.
-trait ExprEnvironment {
-    type Input;
-
-    fn use_database_expr(
-        &mut self,
-        database: &DbInfo,
-        expression: SqlExpr,
-    ) -> anyhow::Result<SqlPageExpr<Self::Input>>;
-}
-
-/// Rejects database-owned inputs because no returned row is available.
-struct StandaloneEnvironment;
-/// Projects database-owned inputs into the current returned row.
-struct RowEnvironment<'a> {
-    row_input_json: &'a mut Vec<bool>,
-    private_projection: &'a mut Vec<SelectItem>,
-}
-
-impl ExprEnvironment for StandaloneEnvironment {
-    type Input = NoRowInput;
-
-    fn use_database_expr(
-        &mut self,
-        _database: &DbInfo,
-        expression: SqlExpr,
-    ) -> anyhow::Result<StandaloneExpr> {
-        if let SqlExpr::Function(function) = &expression
-            && let [ObjectNamePart::Identifier(name)] = function.name.0.as_slice()
-        {
-            return Err(anyhow!(
-                "{} is not a supported sqlpage function and cannot be evaluated before the query",
-                name.value
-            ));
-        }
-        Err(anyhow!(
-            "{expression} is a database expression, but its value is required before the query can run"
-        ))
-    }
-}
-
-impl ExprEnvironment for RowEnvironment<'_> {
-    type Input = RowInputId;
-
-    fn use_database_expr(
-        &mut self,
-        _database: &DbInfo,
-        expression: SqlExpr,
-    ) -> anyhow::Result<RowExpr> {
-        let decode_as_json = is_json_expression(&expression);
-        let index = self.row_input_json.len();
-        let name = format!("{SQLPAGE_INPUT_PREFIX}{index}");
-        self.private_projection.push(SelectItem::ExprWithAlias {
-            expr: expression,
-            alias: Ident::with_quote('"', name),
-        });
-        self.row_input_json.push(decode_as_json);
-        Ok(SqlPageExpr::Input(RowInputId::new(index)))
-    }
-}
-
-#[derive(Clone, Copy)]
-/// SQL operations whose semantics are implemented by the shared `SQLPage`
-/// expression evaluator.
-enum EmulatedFunction {
-    Concat,
-    Coalesce,
-    JsonObject,
-    JsonArray,
-}
-
-/// Rewrites one parsed statement into database SQL plus the `SQLPage`
-/// expressions evaluated around it.
-pub(super) fn rewrite_query(
-    mut statement: SqlStatement,
-    database: &DbInfo,
-    semicolon: bool,
-) -> anyhow::Result<Query> {
-    let source_span = source_span(&statement);
-    let mut partitioner = QueryPartitioner {
-        database,
-        row_input_json: Vec::new(),
-        private_projection: Vec::new(),
-    };
-    if let Some(single_row) = rewrite_single_row(&mut statement, &mut partitioner)? {
-        return Ok(Query {
-            body: QueryBody::SingleRow(single_row),
-            source_span,
-        });
-    }
-    let mut lowerer = DatabaseLowerer::new(database);
-    let computed_columns =
-        rewrite_top_level_projection(&mut statement, &mut partitioner, &mut lowerer)?;
-
-    if let SqlStatement::Query(query) = &mut statement
-        && let SetExpr::Select(select) = query.body.as_mut()
-        && select.projection.is_empty()
-    {
-        select.projection.push(SelectItem::ExprWithAlias {
-            expr: SqlExpr::value(Value::Null),
-            alias: Ident::with_quote('"', format!("{SQLPAGE_INPUT_PREFIX}anchor")),
-        });
-        partitioner.row_input_json.push(false);
-    }
-
-    let json_columns = extract_json_columns(&statement, database.database_type)
-        .into_iter()
-        .filter(|name| !name.starts_with(SQLPAGE_INPUT_PREFIX))
-        .collect();
-    lowerer.lower_statement(&mut statement)?;
-    let bindings = lowerer.finish_bindings(&mut statement)?;
-    let sql = format!(
-        "{statement}{semicolon}",
-        semicolon = if semicolon { ";" } else { "" }
-    );
-
-    Ok(Query {
-        body: QueryBody::Database(DatabaseQuery {
-            sql,
-            bindings,
-            row_input_json: partitioner.row_input_json.into_boxed_slice(),
-            computed_columns: computed_columns.into_boxed_slice(),
-            json_columns,
-        }),
-        source_span,
-    })
-}
-
-/// Removes SQLPage-owned projection expressions from the database projection
-/// and appends their private database inputs as a trailing suffix.
-fn rewrite_top_level_projection(
-    statement: &mut SqlStatement,
-    partitioner: &mut QueryPartitioner<'_>,
-    lowerer: &mut DatabaseLowerer<'_>,
-) -> anyhow::Result<Vec<OutputColumn<RowExpr>>> {
-    let SqlStatement::Query(query) = statement else {
-        return Ok(Vec::new());
-    };
-    let SetExpr::Select(select) = query.body.as_mut() else {
-        return Ok(Vec::new());
-    };
-    if select.distinct.is_some() && select.projection.iter().any(select_item_contains_sqlpage) {
-        anyhow::bail!(
-            "SQLPage-computed projections cannot be used with SELECT DISTINCT because DISTINCT must be evaluated by the database"
-        );
-    }
-
-    let (mut database_projection, computed_columns) =
-        rewrite_projection_items(std::mem::take(&mut select.projection), partitioner, lowerer)?;
-    database_projection.append(&mut partitioner.private_projection);
-    select.projection = database_projection;
-
-    reject_computed_alias_references("WHERE", select.selection.as_ref(), &computed_columns)?;
-    reject_computed_group_by_references(&select.group_by, &computed_columns)?;
-    reject_computed_alias_references("HAVING", select.having.as_ref(), &computed_columns)?;
-    reject_computed_alias_references("QUALIFY", select.qualify.as_ref(), &computed_columns)?;
-    reject_computed_aliases_in_expressions("CLUSTER BY", &select.cluster_by, &computed_columns)?;
-    reject_computed_aliases_in_expressions(
-        "DISTRIBUTE BY",
-        &select.distribute_by,
-        &computed_columns,
-    )?;
-    reject_computed_ordering_references("SORT BY", &select.sort_by, &computed_columns)?;
-    reject_computed_order_by_references(query.order_by.as_ref(), &computed_columns)?;
-    Ok(computed_columns)
-}
-
-fn rewrite_projection_items(
-    projection: Vec<SelectItem>,
-    partitioner: &mut QueryPartitioner<'_>,
-    lowerer: &mut DatabaseLowerer<'_>,
-) -> anyhow::Result<(Vec<SelectItem>, Vec<OutputColumn<RowExpr>>)> {
-    let mut database_projection = Vec::with_capacity(projection.len());
-    let mut computed_columns = Vec::new();
-    for item in projection {
-        let (expression, alias) = match item {
-            SelectItem::ExprWithAlias { expr, alias } => (expr, Some(alias)),
-            SelectItem::UnnamedExpr(expr) => (expr, None),
-            item => {
-                database_projection.push(item);
-                continue;
-            }
-        };
-        let name = alias
-            .as_ref()
-            .map_or_else(|| expression.to_string(), |alias| alias.value.clone());
-        let first_private_input = partitioner.private_projection.len();
-        let projection = partitioner.rewrite_projection(expression)?;
-        match lowerer.lower_projection(
-            projection,
-            &mut partitioner.private_projection[first_private_input..],
-        )? {
-            LoweredProjection::Database(expression) => {
-                database_projection.push(match alias {
-                    Some(alias) => SelectItem::ExprWithAlias {
-                        expr: expression,
-                        alias,
-                    },
-                    None => SelectItem::UnnamedExpr(expression),
-                });
-            }
-            LoweredProjection::PerRow(value) => {
-                computed_columns.push(OutputColumn { name, value });
-            }
-        }
-    }
-    Ok((database_projection, computed_columns))
-}
-
-fn reject_computed_order_by_references(
-    order_by: Option<&sqlparser::ast::OrderBy>,
-    computed_columns: &[OutputColumn<RowExpr>],
-) -> anyhow::Result<()> {
-    if let Some(order_by) = order_by
-        && let OrderByKind::Expressions(expressions) = &order_by.kind
-    {
-        reject_computed_ordering_references("ORDER BY", expressions, computed_columns)?;
-    }
-    Ok(())
-}
-
-fn reject_computed_ordering_references(
-    clause: &str,
-    expressions: &[sqlparser::ast::OrderByExpr],
-    computed_columns: &[OutputColumn<RowExpr>],
-) -> anyhow::Result<()> {
-    if expressions
-        .iter()
-        .any(|ordering| references_computed_projection(&ordering.expr, computed_columns, true))
-    {
-        anyhow::bail!(
-            "{clause} cannot reference a SQLPage-computed column because ordering is performed by the database"
-        );
-    }
-    Ok(())
-}
-
-fn reject_computed_group_by_references(
-    group_by: &GroupByExpr,
-    computed_columns: &[OutputColumn<RowExpr>],
-) -> anyhow::Result<()> {
-    let GroupByExpr::Expressions(expressions, _) = group_by else {
-        return Ok(());
-    };
-    if expressions
-        .iter()
-        .any(|expression| references_computed_projection(expression, computed_columns, true))
-    {
-        anyhow::bail!(
-            "GROUP BY cannot reference a SQLPage-computed column because grouping is performed by the database"
-        );
-    }
-    Ok(())
-}
-
-fn reject_computed_alias_references(
-    clause: &str,
-    expression: Option<&SqlExpr>,
-    computed_columns: &[OutputColumn<RowExpr>],
-) -> anyhow::Result<()> {
-    if expression.is_some_and(|expression| {
-        references_computed_projection(expression, computed_columns, false)
-    }) {
-        anyhow::bail!(
-            "{clause} cannot reference a SQLPage-computed column because it is evaluated by the database"
-        );
-    }
-    Ok(())
-}
-
-fn reject_computed_aliases_in_expressions(
-    clause: &str,
-    expressions: &[SqlExpr],
-    computed_columns: &[OutputColumn<RowExpr>],
-) -> anyhow::Result<()> {
-    if expressions
-        .iter()
-        .any(|expression| references_computed_projection(expression, computed_columns, false))
-    {
-        anyhow::bail!(
-            "{clause} cannot reference a SQLPage-computed column because it is evaluated by the database"
-        );
-    }
-    Ok(())
-}
-
-fn references_computed_projection(
-    expression: &SqlExpr,
-    computed_columns: &[OutputColumn<RowExpr>],
-    reject_ordinal: bool,
-) -> bool {
-    if computed_columns.is_empty() {
-        return false;
-    }
-    if reject_ordinal
-        && matches!(
-            expression,
-            SqlExpr::Value(ValueWithSpan {
-                value: Value::Number(_, _),
-                ..
-            })
-        )
-    {
-        return true;
-    }
-
-    let mut finder = ComputedAliasFinder { computed_columns };
-    sqlparser::ast::Visit::visit(expression, &mut finder).is_break()
-}
-
-/// Rewrites a guaranteed one-row query directly as standalone expressions,
-/// avoiding both a database round trip and an intermediate row-expression tree.
-fn rewrite_single_row(
-    statement: &mut SqlStatement,
-    partitioner: &mut QueryPartitioner<'_>,
-) -> anyhow::Result<Option<SingleRowQuery>> {
-    if !has_single_row_shape(statement) {
-        return Ok(None);
-    }
-    let SqlStatement::Query(query) = statement else {
-        return Ok(None);
-    };
-    let SetExpr::Select(select) = query.body.as_mut() else {
-        return Ok(None);
-    };
-    for item in &select.projection {
-        let SelectItem::ExprWithAlias { expr, .. } = item else {
-            return Ok(None);
-        };
-        if !can_build_standalone(expr)? {
-            return Ok(None);
-        }
-    }
-
-    let mut columns = Vec::with_capacity(select.projection.len());
-    for item in std::mem::take(&mut select.projection) {
-        let SelectItem::ExprWithAlias { expr, alias } = item else {
-            unreachable!("projection shape was checked")
-        };
-        columns.push(OutputColumn {
-            name: alias.value,
-            value: build_sqlpage_expr(partitioner.database, &mut StandaloneEnvironment, expr)?,
-        });
-    }
-    Ok(Some(SingleRowQuery {
-        columns: columns.into_boxed_slice(),
-    }))
-}
-
-fn has_single_row_shape(statement: &SqlStatement) -> bool {
-    let SqlStatement::Query(query) = statement else {
-        return false;
-    };
-    if query.with.is_some()
-        || query.order_by.is_some()
-        || query.limit_clause.is_some()
-        || query.fetch.is_some()
-        || !query.locks.is_empty()
-        || query.for_clause.is_some()
-        || query.settings.is_some()
-        || query.format_clause.is_some()
-        || !query.pipe_operators.is_empty()
-    {
-        return false;
-    }
-    let SetExpr::Select(select) = query.body.as_ref() else {
-        return false;
-    };
-    select.distinct.is_none()
-        && select.top.is_none()
-        && select.into.is_none()
-        && select.from.is_empty()
-        && select.lateral_views.is_empty()
-        && select.selection.is_none()
-        && select.group_by == GroupByExpr::Expressions(vec![], vec![])
-        && select.cluster_by.is_empty()
-        && select.distribute_by.is_empty()
-        && select.sort_by.is_empty()
-        && select.having.is_none()
-        && select.named_window.is_empty()
-        && select.qualify.is_none()
-        && select.prewhere.is_none()
-        && select.connect_by.is_empty()
-        && select.optimizer_hints.is_empty()
-        && select.select_modifiers.is_none()
-        && select.exclude.is_none()
-}
-
-/// Checks standalone support without consuming or cloning the AST, allowing
-/// callers to select a rewrite path before moving any nodes.
-fn can_build_standalone(expression: &SqlExpr) -> anyhow::Result<bool> {
-    match expression {
-        SqlExpr::Value(ValueWithSpan {
-            value:
-                Value::Boolean(_)
-                | Value::Number(_, _)
-                | Value::SingleQuotedString(_)
-                | Value::Null
-                | Value::Placeholder(_),
-            ..
-        }) => Ok(true),
-        SqlExpr::Identifier(identifier) => Ok(variable_from_ident(identifier).is_some()),
-        SqlExpr::Function(function) => {
-            if recognize_sqlpage_function(function)?.is_none()
-                && emulated_function(function).is_none()
-            {
-                return Ok(false);
-            }
-            let FunctionArguments::List(arguments) = &function.args else {
-                return Ok(false);
-            };
-            for argument in &arguments.args {
-                let FunctionArg::Unnamed(FunctionArgExpr::Expr(expression)) = argument else {
-                    return Ok(false);
-                };
-                if !can_build_standalone(expression)? {
-                    return Ok(false);
-                }
-            }
-            Ok(true)
-        }
-        SqlExpr::BinaryOp {
-            left,
-            op: BinaryOperator::StringConcat,
-            right,
-        } => Ok(can_build_standalone(left)? && can_build_standalone(right)?),
-        SqlExpr::Nested(expression) => can_build_standalone(expression),
-        _ => Ok(false),
-    }
-}
-
-impl QueryPartitioner<'_> {
-    /// Splits a projected expression at SQLPage-supported operations while
-    /// leaving opaque database operations in the SQL AST.
-    fn rewrite_projection(&mut self, expression: SqlExpr) -> anyhow::Result<PartitionedProjection> {
-        if !projection_is_per_row(&expression)? {
-            return Ok(PartitionedProjection::Database(expression));
-        }
-        match expression {
-            SqlExpr::Function(function) => {
-                if recognize_sqlpage_function(&function)?.is_some() {
-                    return self
-                        .build_row_expr(SqlExpr::Function(function))
-                        .map(PartitionedProjection::PerRow);
-                }
-                let kind = emulated_function(&function)
-                    .expect("per-row function ownership was already classified");
-                self.rewrite_emulated_projection(function, kind)
-            }
-            SqlExpr::BinaryOp {
-                left,
-                op: BinaryOperator::StringConcat,
-                right,
-            } => {
-                let left = self.rewrite_projection(*left)?;
-                let left = self.projection_into_row_expr(left)?;
-                let right = self.rewrite_projection(*right)?;
-                let right = self.projection_into_row_expr(right)?;
-                Ok(PartitionedProjection::PerRow(SqlPageExpr::Concat {
-                    arguments: vec![left, right].into_boxed_slice(),
-                    null_behavior: self.database.database_type.concat_operator_null_behavior(),
-                }))
-            }
-            SqlExpr::Nested(expression) => match self.rewrite_projection(*expression)? {
-                PartitionedProjection::Database(expression) => Ok(PartitionedProjection::Database(
-                    SqlExpr::Nested(Box::new(expression)),
-                )),
-                PartitionedProjection::PerRow(expression) => {
-                    Ok(PartitionedProjection::PerRow(expression))
-                }
-            },
-            _ => unreachable!("per-row expression ownership was already classified"),
-        }
-    }
-
-    fn rewrite_emulated_projection(
-        &mut self,
-        function: Function,
-        kind: EmulatedFunction,
-    ) -> anyhow::Result<PartitionedProjection> {
-        let (arguments, _) = take_expression_arguments(function)?;
-        let mut rewritten = Vec::with_capacity(arguments.len());
-        for argument in arguments {
-            let argument = self.rewrite_projection(argument)?;
-            rewritten.push(self.projection_into_row_expr(argument)?);
-        }
-        Ok(PartitionedProjection::PerRow(build_emulated(
-            kind,
-            rewritten,
-            self.database.database_type,
-        )?))
-    }
-
-    /// Converts a database-owned projection fragment into a typed row input,
-    /// while preserving an already per-row expression unchanged.
-    fn projection_into_row_expr(
-        &mut self,
-        projection: PartitionedProjection,
-    ) -> anyhow::Result<RowExpr> {
-        match projection {
-            PartitionedProjection::Database(expression) => self.build_row_expr(expression),
-            PartitionedProjection::PerRow(expression) => Ok(expression),
-        }
-    }
-
-    fn build_row_expr(&mut self, expression: SqlExpr) -> anyhow::Result<RowExpr> {
-        build_sqlpage_expr(
-            self.database,
-            &mut RowEnvironment {
-                row_input_json: &mut self.row_input_json,
-                private_projection: &mut self.private_projection,
-            },
-            expression,
-        )
-    }
-}
-
-/// Determines ownership without mutating the AST, so database fragments can
-/// be promoted to row inputs in source order once a per-row parent is known.
-fn projection_is_per_row(expression: &SqlExpr) -> anyhow::Result<bool> {
-    match expression {
-        SqlExpr::Function(function) => {
-            if recognize_sqlpage_function(function)?.is_some() {
-                return Ok(true);
-            }
-            if emulated_function(function).is_none() {
-                return Ok(false);
-            }
-            let FunctionArguments::List(arguments) = &function.args else {
-                anyhow::bail!("Unsupported arguments to {}", function.name);
-            };
-            if arguments.duplicate_treatment.is_some() || !arguments.clauses.is_empty() {
-                anyhow::bail!("Unsupported arguments to {}", function.name);
-            }
-            for argument in &arguments.args {
-                let FunctionArg::Unnamed(FunctionArgExpr::Expr(expression)) = argument else {
-                    anyhow::bail!("Named and wildcard function arguments are not supported");
-                };
-                if projection_is_per_row(expression)? {
-                    return Ok(true);
-                }
-            }
-            Ok(false)
-        }
-        SqlExpr::BinaryOp {
-            left,
-            op: BinaryOperator::StringConcat,
-            right,
-        } => Ok(projection_is_per_row(left)? || projection_is_per_row(right)?),
-        SqlExpr::Nested(expression) => projection_is_per_row(expression),
-        _ => Ok(false),
-    }
-}
-
-impl<'a> DatabaseLowerer<'a> {
-    fn new(database: &'a DbInfo) -> Self {
-        Self {
-            database,
-            bindings: Vec::new(),
-            error: None,
-        }
-    }
-
-    /// Accepts a terminal partition result and lowers its database-owned
-    /// fragments. The lowered AST cannot return to recursive partitioning.
-    fn lower_projection(
-        &mut self,
-        projection: PartitionedProjection,
-        private_inputs: &mut [SelectItem],
-    ) -> anyhow::Result<LoweredProjection> {
-        let projection = match projection {
-            PartitionedProjection::Database(mut expression) => {
-                self.lower_expression(&mut expression)?;
-                LoweredProjection::Database(expression)
-            }
-            PartitionedProjection::PerRow(expression) => {
-                for input in private_inputs {
-                    let _ = input.visit(self);
-                    self.take_error()?;
-                }
-                LoweredProjection::PerRow(expression)
-            }
-        };
-        Ok(projection)
-    }
-
-    fn lower_expression(&mut self, expression: &mut SqlExpr) -> anyhow::Result<()> {
-        let _ = expression.visit(self);
-        self.take_error()
-    }
-
-    fn lower_statement(&mut self, statement: &mut SqlStatement) -> anyhow::Result<()> {
-        let _ = statement.visit(self);
-        self.take_error()
-    }
-
-    fn take_error(&mut self) -> anyhow::Result<()> {
-        self.error.take().map_or(Ok(()), Err)
-    }
-
-    fn add_binding(&mut self, value: StandaloneExpr) -> SqlExpr {
-        let sequence = self.bindings.len();
-        self.bindings.push(value);
-        let placeholder = match placeholder_style(self.database.kind) {
-            PlaceholderStyle::Numbered { prefix } => format!("{prefix}{}", sequence + 1),
-            PlaceholderStyle::Positional { .. } => format!("${}", sequence + 1),
-        };
-        cast_placeholder(placeholder, self.database)
-    }
-
-    fn finish_bindings(
-        &mut self,
-        statement: &mut SqlStatement,
-    ) -> anyhow::Result<Box<[StandaloneExpr]>> {
-        let PlaceholderStyle::Positional { token } = placeholder_style(self.database.kind) else {
-            return Ok(std::mem::take(&mut self.bindings).into_boxed_slice());
-        };
-        let mut finalizer = PositionalBindingFinalizer {
-            token,
-            binding_count: self.bindings.len(),
-            order: Vec::with_capacity(self.bindings.len()),
-            error: None,
-        };
-        let _ = statement.visit(&mut finalizer);
-        if let Some(error) = finalizer.error {
-            return Err(error);
-        }
-        let mut bindings = std::mem::take(&mut self.bindings)
-            .into_iter()
-            .map(Some)
-            .collect::<Vec<_>>();
-        finalizer
-            .order
-            .into_iter()
-            .map(|index| {
-                bindings[index].take().ok_or_else(|| {
-                    anyhow!("Generated binding placeholder ${} was repeated", index + 1)
-                })
-            })
-            .collect::<anyhow::Result<Box<_>>>()
-    }
-}
-
-struct PositionalBindingFinalizer {
-    token: &'static str,
-    binding_count: usize,
-    order: Vec<usize>,
-    error: Option<anyhow::Error>,
-}
-
-impl VisitorMut for PositionalBindingFinalizer {
-    type Break = ();
-
-    fn pre_visit_expr(&mut self, expression: &mut SqlExpr) -> ControlFlow<Self::Break> {
-        let SqlExpr::Value(ValueWithSpan {
-            value: Value::Placeholder(name),
-            span,
-        }) = expression
-        else {
-            return ControlFlow::Continue(());
-        };
-        if *span != Span::empty() {
-            return ControlFlow::Continue(());
-        }
-        let Some(index) = name
-            .strip_prefix('$')
-            .and_then(|number| number.parse::<usize>().ok())
-            .and_then(|number| number.checked_sub(1))
-        else {
-            return ControlFlow::Continue(());
-        };
-        if index >= self.binding_count {
-            self.error = Some(anyhow!("Invalid generated binding placeholder {name}"));
-            return ControlFlow::Break(());
-        }
-        self.order.push(index);
-        self.token.clone_into(name);
-        ControlFlow::Continue(())
-    }
-}
-
-impl VisitorMut for DatabaseLowerer<'_> {
-    type Break = ();
-
-    fn pre_visit_expr(&mut self, expression: &mut SqlExpr) -> ControlFlow<Self::Break> {
-        if self.error.is_some() {
-            return ControlFlow::Break(());
-        }
-
-        let replacement = match expression {
-            SqlExpr::Value(ValueWithSpan {
-                value: Value::Placeholder(_),
-                span,
-            }) if *span == Span::empty() => None,
-            SqlExpr::Value(ValueWithSpan {
-                value: Value::Placeholder(_),
-                ..
-            })
-            | SqlExpr::Identifier(_) => variable_from_expr(expression)
-                .map(|variable| self.add_binding(SqlPageExpr::Variable(variable))),
-            SqlExpr::Function(function) => match recognize_sqlpage_function(function) {
-                Ok(Some(_)) => {
-                    let owned = std::mem::replace(expression, SqlExpr::value(Value::Null));
-                    match build_sqlpage_expr(self.database, &mut StandaloneEnvironment, owned) {
-                        Ok(value) => Some(self.add_binding(value)),
-                        Err(error) => {
-                            self.error = Some(error.context(
-                                "A SQLPage function used by the database could not be evaluated before the query",
-                            ));
-                            None
-                        }
-                    }
-                }
-                Ok(None) => None,
-                Err(error) => {
-                    self.error = Some(error);
-                    None
-                }
-            },
-            SqlExpr::BinaryOp {
-                left,
-                op: BinaryOperator::StringConcat,
-                right,
-            } if self.database.database_type == SupportedDatabase::Mssql => {
-                let left = std::mem::replace(left.as_mut(), SqlExpr::value(Value::Null));
-                let right = std::mem::replace(right.as_mut(), SqlExpr::value(Value::Null));
-                Some(make_function("CONCAT", vec![left, right]))
-            }
-            SqlExpr::Cast {
-                kind: kind @ CastKind::DoubleColon,
-                ..
-            } if ![
-                SupportedDatabase::Postgres,
-                SupportedDatabase::Duckdb,
-                SupportedDatabase::Snowflake,
-                SupportedDatabase::Generic,
-            ]
-            .contains(&self.database.database_type) =>
-            {
-                *kind = CastKind::Cast;
-                None
-            }
-            _ => None,
-        };
-
-        if let Some(replacement) = replacement {
-            *expression = replacement;
-        }
-        ControlFlow::Continue(())
-    }
-}
-
-/// Consumes an AST expression into the shared `SQLPage` expression type. The
-/// environment determines whether opaque database fragments are illegal or
-/// become private row inputs.
-fn build_sqlpage_expr<Environment: ExprEnvironment>(
-    database: &DbInfo,
-    environment: &mut Environment,
-    expression: SqlExpr,
-) -> anyhow::Result<SqlPageExpr<Environment::Input>> {
-    match expression {
-        SqlExpr::Value(ValueWithSpan { value, .. }) => match value {
-            Value::Placeholder(name) => Ok(SqlPageExpr::Variable(variable_from_placeholder(name))),
-            Value::SingleQuotedString(text) => Ok(SqlPageExpr::Literal(JsonValue::String(text))),
-            Value::Number(number, _) => Ok(SqlPageExpr::Literal(JsonValue::Number(
-                number.parse().context("Invalid numeric SQL literal")?,
-            ))),
-            Value::Boolean(value) => Ok(SqlPageExpr::Literal(JsonValue::Bool(value))),
-            Value::Null => Ok(SqlPageExpr::Literal(JsonValue::Null)),
-            _ => {
-                environment.use_database_expr(database, SqlExpr::Value(ValueWithSpan::from(value)))
-            }
-        },
-        SqlExpr::Identifier(identifier) => {
-            if let Some(variable) = variable_from_ident(&identifier) {
-                Ok(SqlPageExpr::Variable(variable))
-            } else {
-                environment.use_database_expr(database, SqlExpr::Identifier(identifier))
-            }
-        }
-        SqlExpr::Function(function) => {
-            if let Some(function_name) = recognize_sqlpage_function(&function)? {
-                let (arguments, _) = take_expression_arguments(function)?;
-                let arguments = arguments
-                    .into_iter()
-                    .map(|argument| build_sqlpage_expr(database, environment, argument))
-                    .collect::<anyhow::Result<Vec<_>>>()?;
-                Ok(SqlPageExpr::Call {
-                    function: function_name,
-                    arguments: arguments.into_boxed_slice(),
-                })
-            } else if let Some(kind) = emulated_function(&function) {
-                let (arguments, _) = take_expression_arguments(function)?;
-                let arguments = arguments
-                    .into_iter()
-                    .map(|argument| build_sqlpage_expr(database, environment, argument))
-                    .collect::<anyhow::Result<Vec<_>>>()?;
-                build_emulated(kind, arguments, database.database_type)
-            } else {
-                environment.use_database_expr(database, SqlExpr::Function(function))
-            }
-        }
-        SqlExpr::BinaryOp {
-            left,
-            op: BinaryOperator::StringConcat,
-            right,
-        } => Ok(SqlPageExpr::Concat {
-            arguments: vec![
-                build_sqlpage_expr(database, environment, *left)?,
-                build_sqlpage_expr(database, environment, *right)?,
-            ]
-            .into_boxed_slice(),
-            null_behavior: database.database_type.concat_operator_null_behavior(),
-        }),
-        SqlExpr::Nested(expression) => build_sqlpage_expr(database, environment, *expression),
-        expression => environment.use_database_expr(database, expression),
-    }
-}
-
-fn build_emulated<Input>(
-    kind: EmulatedFunction,
-    arguments: Vec<SqlPageExpr<Input>>,
-    database: SupportedDatabase,
-) -> anyhow::Result<SqlPageExpr<Input>> {
-    Ok(match kind {
-        EmulatedFunction::Concat => SqlPageExpr::Concat {
-            arguments: arguments.into_boxed_slice(),
-            null_behavior: database.concat_function_null_behavior(),
-        },
-        EmulatedFunction::Coalesce => SqlPageExpr::Coalesce(arguments.into_boxed_slice()),
-        EmulatedFunction::JsonArray => SqlPageExpr::JsonArray(arguments.into_boxed_slice()),
-        EmulatedFunction::JsonObject => {
-            if !arguments.len().is_multiple_of(2) {
-                anyhow::bail!("JSON_OBJECT requires an even number of arguments");
-            }
-            let mut arguments = arguments.into_iter();
-            let mut entries = Vec::with_capacity(arguments.len() / 2);
-            while let Some(key) = arguments.next() {
-                let value = arguments.next().expect("argument count was checked");
-                entries.push((key, value));
-            }
-            SqlPageExpr::JsonObject(entries.into_boxed_slice())
-        }
-    })
-}
-
-/// Recognizes and validates an unquoted `sqlpage.<name>` call. A recognized
-/// call is either rewritten or rejected and can never reach database SQL.
-fn recognize_sqlpage_function(function: &Function) -> anyhow::Result<Option<SqlPageFunctionName>> {
-    let ObjectName(parts) = &function.name;
-    if !is_sqlpage_func(parts) {
-        return Ok(None);
-    }
-    let [
-        ObjectNamePart::Identifier(_),
-        ObjectNamePart::Identifier(name),
-    ] = parts.as_slice()
-    else {
-        unreachable!("is_sqlpage_func checked the name")
-    };
-    if function.uses_odbc_syntax
-        || !matches!(function.parameters, FunctionArguments::None)
-        || function.filter.is_some()
-        || function.null_treatment.is_some()
-        || function.over.is_some()
-        || !function.within_group.is_empty()
-    {
-        anyhow::bail!(
-            "Modifiers are not supported on SQLPage function {}",
-            function.name
-        );
-    }
-    let FunctionArguments::List(FunctionArgumentList {
-        duplicate_treatment: None,
-        clauses,
-        ..
-    }) = &function.args
-    else {
-        anyhow::bail!(
-            "Unsupported argument syntax for SQLPage function {}",
-            function.name
-        );
-    };
-    if !clauses.is_empty() {
-        anyhow::bail!(
-            "Argument clauses are not supported on SQLPage function {}",
-            function.name
-        );
-    }
-    Ok(Some(SqlPageFunctionName::from_str(&name.value)?))
-}
-
-fn emulated_function(function: &Function) -> Option<EmulatedFunction> {
-    let [ObjectNamePart::Identifier(name)] = function.name.0.as_slice() else {
-        return None;
-    };
-    if !matches!(function.parameters, FunctionArguments::None)
-        || function.filter.is_some()
-        || function.null_treatment.is_some()
-        || function.over.is_some()
-        || !function.within_group.is_empty()
-    {
-        return None;
-    }
-    match name.value.to_ascii_lowercase().as_str() {
-        "concat" => Some(EmulatedFunction::Concat),
-        "coalesce" => Some(EmulatedFunction::Coalesce),
-        "json_object" | "jsonb_object" | "json_build_object" | "jsonb_build_object" => {
-            Some(EmulatedFunction::JsonObject)
-        }
-        "json_array" | "jsonb_array" | "json_build_array" | "jsonb_build_array" => {
-            Some(EmulatedFunction::JsonArray)
-        }
-        _ => None,
-    }
-}
-
-/// Moves expression arguments out of a function while retaining its emptied
-/// AST shell so database-owned functions can be rebuilt without cloning.
-fn take_expression_arguments(mut function: Function) -> anyhow::Result<(Vec<SqlExpr>, Function)> {
-    let FunctionArguments::List(arguments) = &mut function.args else {
-        anyhow::bail!("Unsupported arguments to {}", function.name);
-    };
-    if arguments.duplicate_treatment.is_some() || !arguments.clauses.is_empty() {
-        anyhow::bail!("Unsupported arguments to {}", function.name);
-    }
-    let arguments = std::mem::take(&mut arguments.args)
-        .into_iter()
-        .map(|argument| match argument {
-            FunctionArg::Unnamed(FunctionArgExpr::Expr(expression)) => Ok(expression),
-            _ => Err(anyhow!(
-                "Named and wildcard function arguments are not supported"
-            )),
-        })
-        .collect::<anyhow::Result<Vec<_>>>()?;
-    Ok((arguments, function))
-}
-
-fn make_function(name: &str, expressions: Vec<SqlExpr>) -> SqlExpr {
-    SqlExpr::Function(Function {
-        name: ObjectName(vec![ObjectNamePart::Identifier(Ident::new(name))]),
-        args: FunctionArguments::List(FunctionArgumentList {
-            args: expressions
-                .into_iter()
-                .map(|expression| FunctionArg::Unnamed(FunctionArgExpr::Expr(expression)))
-                .collect(),
-            duplicate_treatment: None,
-            clauses: Vec::new(),
-        }),
-        parameters: FunctionArguments::None,
-        over: None,
-        filter: None,
-        null_treatment: None,
-        within_group: Vec::new(),
-        uses_odbc_syntax: false,
-    })
-}
-
-fn variable_from_expr(expression: &SqlExpr) -> Option<VariableRef> {
-    match expression {
-        SqlExpr::Value(ValueWithSpan {
-            value: Value::Placeholder(name),
-            ..
-        }) => Some(variable_from_placeholder(name.clone())),
-        SqlExpr::Identifier(identifier) => variable_from_ident(identifier),
-        _ => None,
-    }
-}
-
-fn variable_from_ident(identifier: &Ident) -> Option<VariableRef> {
-    if identifier.quote_style.is_some() {
-        return None;
-    }
-    let prefix = identifier.value.chars().next()?;
-    matches!(prefix, '$' | ':' | '?').then(|| VariableRef {
-        name: identifier.value[prefix.len_utf8()..].to_owned(),
-        source: variable_source(prefix),
-    })
-}
-
-fn variable_from_placeholder(mut name: String) -> VariableRef {
-    let prefix = name.remove(0);
-    VariableRef {
-        name,
-        source: variable_source(prefix),
-    }
-}
-
-fn variable_source(prefix: char) -> VariableSource {
-    match prefix {
-        '$' => VariableSource::SetOrUrl,
-        ':' => VariableSource::SetOrForm,
-        _ => VariableSource::Url,
-    }
-}
-
-/// Wraps a generated placeholder in the backend-specific text cast when the
-/// database cannot reliably infer that the parameter is a string.
-///
-/// `SQLPage` always binds parameters as strings. Native `PostgreSQL` (which
-/// pins the parameter type to `TEXT` when preparing the statement), `MySQL`
-/// and `SQL Server` (which convert the bound string to the type expected by
-/// the surrounding expression) do not need the cast, and it can even be
-/// harmful: on `SQL Server` the parameter is bound as `NVARCHAR(MAX)`, and
-/// casting it to a narrow `VARCHAR` mangles non-ASCII values. `SQLite`
-/// needs it to keep text affinity in comparisons with numbers.
-///
-/// Through ODBC, the decision follows the database behind the driver, since
-/// `SQLPage` knows it from the driver's reported name:
-/// - `PostgreSQL` keeps the cast: `psqlodbc` provides no parameter type
-///   information, and the server then fails on context-free parameters
-///   (`could not determine data type of parameter`).
-/// - `SQLite` keeps it for the same affinity reasons as native connections.
-/// - `MySQL`, `SQL Server` and `DuckDB` drop it, like their native
-///   counterparts: the former two convert the string at execution time, and
-///   `DuckDB` defaults untyped parameters to `VARCHAR`.
-/// - `Oracle`, `Snowflake` and unknown databases keep it conservatively.
-fn cast_placeholder(placeholder: String, database: &DbInfo) -> SqlExpr {
-    let data_type = match database.kind {
-        AnyKind::Sqlite => DataType::Text,
-        AnyKind::Postgres | AnyKind::MySql | AnyKind::Mssql => {
-            return SqlExpr::value(Value::Placeholder(placeholder));
-        }
-        AnyKind::Odbc => match database.database_type {
-            SupportedDatabase::Postgres | SupportedDatabase::Sqlite => DataType::Text,
-            SupportedDatabase::Oracle => DataType::Varchar(Some(CharacterLength::IntegerLength {
-                length: 4000,
-                unit: None,
-            })),
-            SupportedDatabase::MySql | SupportedDatabase::Mssql | SupportedDatabase::Duckdb => {
-                return SqlExpr::value(Value::Placeholder(placeholder));
-            }
-            _ => DataType::Varchar(None),
-        },
-    };
-    SqlExpr::Cast {
-        expr: Box::new(SqlExpr::value(Value::Placeholder(placeholder))),
-        data_type,
-        format: None,
-        kind: CastKind::Cast,
-        array: false,
-    }
-}
-
-fn source_span(value: &impl sqlparser::ast::Spanned) -> SourceSpan {
-    let span = value.span();
-    SourceSpan {
-        start: SourceLocation {
-            line: usize::try_from(span.start.line).unwrap_or(0),
-            column: usize::try_from(span.start.column).unwrap_or(0),
-        },
-        end: SourceLocation {
-            line: usize::try_from(span.end.line).unwrap_or(0),
-            column: usize::try_from(span.end.column).unwrap_or(0),
-        },
-    }
-}
-
-fn select_item_contains_sqlpage(item: &SelectItem) -> bool {
-    struct Finder(bool);
-    impl sqlparser::ast::Visitor for Finder {
-        type Break = ();
-
-        fn pre_visit_expr(&mut self, expression: &SqlExpr) -> ControlFlow<Self::Break> {
-            if let SqlExpr::Function(function) = expression
-                && is_sqlpage_func(&function.name.0)
-            {
-                self.0 = true;
-                return ControlFlow::Break(());
-            }
-            ControlFlow::Continue(())
-        }
-    }
-    let mut finder = Finder(false);
-    let _ = sqlparser::ast::Visit::visit(item, &mut finder);
-    finder.0
-}
